Add files via upload
This project was a cool little side adventure where I got to test out my coding skills and have a blast with my friends. I also threw in a touchscreen feature, which was a neat little addition to my very own Flappy Bird Game. To top it off, I used an online converter to turn it into a playable app, making it super easy to share with my friends. 

Overall, the project was about more than just coding. It was about creating something enjoyable and sharing it with others. Playing the game with my friends and seeing them have fun with something I created has been one of the best parts. 